requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.4
flake8==6.0.0
flake8-html==0.4.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.4.0
pytz==2022.6
sqlparse==0.4.3
